Fix #123 by moving polyfill bindings to root module
I beleive the issue is that we attached the polyfills in FirebaseActions but
they are necessary in FirebaseConnector. Once moving the polyfill binding to
index.commo I was not able to reproduce this issue.

See this google group post for more write up: https://groups.google.com/forum/#!topic/firebase-talk/ejc4hbwUNy4